% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_measurement_schedule.R
\name{compute_measurement_schedule}
\alias{compute_measurement_schedule}
\title{Computes measurement schedules.}
\usage{
compute_measurement_schedule(
  time_period,
  num_measurements,
  smallest_int_length,
  measurement_spacing
)
}
\arguments{
\item{time_period}{number of days over which change occurs}

\item{num_measurements}{number of measurements}

\item{smallest_int_length}{length of smallest intervals}

\item{measurement_spacing}{type of measurement spacing (one of 'equal', 'time_inc', 'time_dec', and 'mid_ext')}
}
\value{
Returns a list.
}
\description{
Computes measurement schedules.
}
