% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logistic_model_ns.R
\name{create_logistic_growth_model_ns}
\alias{create_logistic_growth_model_ns}
\title{Analyzes data with nonlinear latent growth curve model.}
\usage{
create_logistic_growth_model_ns(data_wide, model_name)
}
\arguments{
\item{data_wide}{wide version of data}

\item{model_name}{name of model}
}
\description{
Data are modelled with a structured latent growth curve model.
}
