% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_delay_table.R
\name{generate_delay_table}
\alias{generate_delay_table}
\title{Generates time structuredness manipulation.}
\usage{
generate_delay_table(delay_ls)
}
\arguments{
\item{num_iterations}{number of iterations}

\item{pop_params}{list of population parameters}

\item{schedule}{measurement schedule}

\item{response_group_size}{sample size}
}
\value{
Returns a data table.
}
\description{
Generates time structuredness manipulation.
}
