% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_four_param_cov_matrix.R
\name{generate_four_param_cov_matrix}
\alias{generate_four_param_cov_matrix}
\title{Generates covariance matrix for nonlinear longitudinal data that follow logistic pattern.}
\usage{
generate_four_param_cov_matrix(num_time_points, pop_param_list)
}
\arguments{
\item{num_time_points}{number of time points}

\item{pop_param_list}{list of population parameters returned from generate_pop_param_list}
}
\value{
Returns a covariance matrix.
}
\description{
Four parameters (fixed effects) are used to characterize the logistic pattern and, due to the hierarchical nature of the to-be-generated data,
each parameter has a corresponding value of variability (i.e., random-effect). The random effects are used to generate the covariance
matrix. Note that correlations between random effects must also be set and that correlations between random effects and error variance
at each time point are set to 0 by default (cor_param_error = 0). Internally, the function also assumes zero-value correlations between
error variances at each time point  The four parameters that characterize the logistic pattern of change take on the
following meanings:
\itemize{
\item diff: different between first and last values (i.e., difference between two plateaus)
\item beta: amount of time to reach midway point (i.e., 50\% of the distance between theta and alpha) from time = 0
\item gamma: amount of time to reach satiation point (i.e., 73\% of distance between theta and alpha) from midpoint
}
}
