% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_four_param_pop_curve.R
\name{generate_four_param_pop_curve}
\alias{generate_four_param_pop_curve}
\title{Generates list containing values for each population-level parameter.}
\usage{
generate_four_param_pop_curve(
  theta_fixed,
  alpha_fixed,
  beta_fixed,
  gamma_fixed,
  sd_theta,
  sd_alpha,
  sd_beta,
  sd_gamma,
  sd_error,
  cor_alpha_theta = 0,
  cor_beta_theta = 0,
  cor_beta_alpha = 0,
  cor_gamma_theta = 0,
  cor_gamma_alpha = 0,
  cor_gamma_beta = 0,
  cor_param_error = 0
)
}
\arguments{
\item{sd_theta}{standard deviation of  first plateau}

\item{sd_alpha}{standard deviation of second plateau}

\item{sd_gamma}{standard deviation of gamma}

\item{sd_error}{standard deviation of errors}

\item{cor_beta_theta}{beta_theta correlation}

\item{cor_beta_alpha}{beta_alpha correlation}

\item{cor_gamma_theta}{gamma_theta correlation}

\item{cor_gamma_alpha}{gamma_alpha correlation}

\item{cor_gamma_beta}{gamma_beta-gamma correlation}

\item{cor_param_error}{parameter-error correlation}

\item{num_time_points}{number of time points}

\item{cor_theta_alpha}{theta_alpha  correlation}

\item{scaling_constant}{constant that scales scores}
}
\value{
Returns a covariance matrix.
}
\description{
Four parameters (fixed effects) are used to characterize the logistic pattern and, due to the hierarchical nature of the to-be-generated data,
each parameter has a corresponding value of variability (i.e., random-effect). The random effects are used to generate the covariance
matrix. Note that correlations between random effects must also be set and that correlations between random effects and error variance
at each time point are set to 0 by default (cor_param_error = 0). Internally, the function also assumes zero-value correlations between
error variances at each time point  The four parameters that characterize the logistic pattern of change take on the
following meanings:
\itemize{
\item theta: starting value (first plateau)
\item alpha: ending value (second plateau)
\item beta: amount of time to reach midway point (i.e., 50\% of the distance between theta and alpha) from time = 0
\item gamma: amount of time to reach satiation point (i.e., 73\% of distance between theta and alpha) from midpoint
}
}
