% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_param_values.R
\name{generate_ind_param_values}
\alias{generate_ind_param_values}
\title{Generates parameter values for each person in a response group.}
\usage{
generate_ind_param_values(
  pop_param_list,
  response_group_size,
  num_time_points,
  cov_matrix
)
}
\arguments{
\item{response_group_size}{response group size}

\item{num_time_points}{number of time points}

\item{cov_matrix}{covariance matrix generated from generate_cov_matrix}

\item{theta_fixed}{fixed-effect value for theta}

\item{alpha_fixed}{fixed-effect value for alpha}

\item{beta_fixed}{fixed-effect value for beta}

\item{gamma_fixed}{fixed-effect value for gamma}
}
\value{
Returns a data table.
}
\description{
The four parameters that characterize the logistic pattern of change take on the
following meanings:
\itemize{
\item diff: different between first and last values (i.e., difference between two plateaus)
\item beta: amount of time to reach midway point (i.e., 50\% of the distance between theta and alpha) from time = 0
\item gamma: amount of time to reach satiation point (i.e., 73\% of distance between theta and alpha) from midpoint
}
}
