% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_exp_simulations.R
\name{run_exp_simulation}
\alias{run_exp_simulation}
\title{Runs Monte Carlo simulations.}
\usage{
run_exp_simulation(factor_list, num_iterations, pop_params, num_cores, seed)
}
\arguments{
\item{num_iterations}{number of iterations}

\item{pop_params}{list of population parameters}

\item{schedule}{measurement schedule}
}
\value{
Returns a data table.
}
\description{
Runs Monte Carlo simulations.
}
